requirements.txt
together==1.5.7
python-dotenv==1.0.0
requests==2.31.0
Pillow>=11.1.0,<12.0.0 